requirements.txt
pandas>=2.0
pyyaml>=6.0
shapely>=2.0
requests>=2.28
tqdm>=4.65
python-dotenv>=1.0
